refactor(local-ci): extract targets.py (R2-1 eighth cut)
Move 3 target-resolution helpers (enabled_targets, parse_targets_arg,
resolve_targets) out of local_ci.py into a sibling targets.py module.
Pure functions, no I/O — they own the precedence order for the
--targets CLI argument and the config-driven default-target list.

- local_ci.py: 10,843 → 10,803 lines (-40 net).
- targets.py: 70 lines.
- Zero in-package dependencies — leaf module.

Tests: 213/213 local_ci tests pass.

Running R2-1 total: 8 modules extracted (state_paths, normalize,
git_helpers, io_utils, footprint, provenance, job_queue, targets).
local_ci.py shrunk 11,285 → 10,803 lines (-482, -4.3%).
